:bug: Fix-Windows-highlighting
change e.metaKey to hasCommandModifier(e) from draft-js
metaKey makes sense for mac Os (command key) but not for windows,
as CTRL is not metaKey. As a result, highlighting wasn't working
on Windows. Draft JS has a built in utility function, providing
a cross-platform fix, tested and accountable to them.